[docs] padding out tag lines for html to match rst
When showing examples of valid tag lines the list shows up in a single line
in the rendered html, and since the @slow keyword turns up twice this does
look odd.  The rst, however, shows these tag lines on separate lines and it
therefore seems that the desired html formatting was actually three separate
tag lines.  This change makes the rendered html match the rst's intention.